fix: assert delivery
