+++ crates/topos-test-sdk/Cargo.toml
@@ -40,6 +40,7 @@ tower.workspace = true
 tokio-util.workspace = true
 tokio.workspace = true
 tracing.workspace = true
+async-stream.workspace = true
 
 [features]
 default = []
